# openapi.yaml
openapi: 3.0.3
info:
    title: Swagger Petstore - OpenAPI 3.0
    version: 1.0.0
servers:
    - url: http://localhost:3000
...
paths:
    /articles:
        post:
            tags:
                - Article
            description: 게시글 생성 API
            operationId: createArticle
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateArticleRequest'
            responses:
                '201':
                    description: 성공
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Article'
                '400':
                    description: Validation failed

        get:
            tags:
                - Article
            description: 게시글 목록 조회 API
            operationId: getArticleList
            parameters:
                - name: cursor
                  in: query
                  required: false
                  schema:
                      type: integer
                      default: 0
                      minimum: 0
                  example: 0
                - name: take
                  in: query
                  required: false
                  schema:
                      type: integer
                      default: 10
                      minimum: 1
                      maximum: 10
                  example: 10
                - name: orderBy
                  in: query
                  required: false
                  schema:
                      type: string
                      enum:
                          - recent
                  example: recent
                - name: word
                  in: query
                  required: false
                  schema:
                      type: string
                  example: '검색어'
            responses:
                '200':
                    description: 성공
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetArticleListResponse'
    /articles/{articleId}:
        get:
            tags:
                - Article
            description: 게시글 조회 API
            operationId: getArticle
            parameters:
                - name: articleId
                  in: path
                  required: true
                  schema:
                      type: integer
                  example: 1
            responses:
                '200':
                    description: 성공
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Article'
                '404':
                    description: Not found

        patch:
            tags:
                - Article
            description: '게시글 수정 API'
            operationId: updateArticle
            parameters:
                - name: articleId
                  in: path
                  required: true
                  schema:
                      type: integer
                  example: 1
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateArticleRequest'
            responses:
                '200':
                    description: 성공
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Article'
                '400':
                    description: Validation failed
                '404':
                    description: Not found

        delete:
            tags:
                - Article
            description: '게시글 삭제 API'
            operationId: deleteArticle
            parameters:
                - name: articleId
                  in: path
                  required: true
                  schema:
                      type: integer
                  example: 1
            responses:
                '204':
                    description: 성공
                '404':
                    description: Not found
...
components:
    schemas:
        CreateArticleRequest:
            type: object
            required:
                - title
                - content
            properties:
                title:
                    type: string
                    example: '게시글 제목'
                content:
                    type: string
                    example: '게시글 내용'
        GetArticleListResponse:
            type: object
            properties:
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/Article'
                nextCursor:
                    type: integer
                    example: 10
                hasMore:
                    type: boolean
                    example: true
        UpdateArticleRequest:
            type: object
            properties:
                title:
                    type: string
                    example: '게시글 제목'
                content:
                    type: string
                    example: '게시글 내용'
...
        Article:
            type: object
            properties:
                id:
                    type: integer
                    example: 1
                title:
                    type: string
                    example: '게시글 제목'
                content:
                    type: string
                    example: '게시글 내용'
                createdAt:
                    type: string
                    format: datetime
                    example: '2024-07-15T00:00:00.000Z'
...